feat: add base promotional advert
add mobile layout

tweak text container size

only show new banner on Base

fix import

update to new state structure
