+++ src/state/user/reducer.test.ts
@@ -7,8 +7,8 @@ import reducer, {
   addSerializedPair,
   addSerializedToken,
   initialState,
+  updateHideBaseWalletBanner,
   updateHideClosedPositions,
-  updateHideUniswapWalletBanner,
   updateSelectedWallet,
   updateUserDeadline,
   updateUserLocale,
@@ -110,10 +110,10 @@ describe('swap reducer', () => {
     })
   })
 
-  describe('updateHideUniswapWalletBanner', () => {
+  describe('updateHideBaseWalletBanner', () => {
     it('updates the hideUniswapWalletBanner', () => {
-      store.dispatch(updateHideUniswapWalletBanner({ hideUniswapWalletBanner: true }))
-      expect(store.getState().hideUniswapWalletBanner).toEqual(true)
+      store.dispatch(updateHideBaseWalletBanner({ hideBaseWalletBanner: true }))
+      expect(store.getState().hideBaseWalletBanner).toEqual(true)
     })
   })
