+++ src/state/user/reducer.ts
@@ -47,7 +47,7 @@ export interface UserState {
 
   timestamp: number
   URLWarningVisible: boolean
-  hideUniswapWalletBanner: boolean
+  hideBaseWalletBanner: boolean
   disabledUniswapX?: boolean
   // undefined means has not gone through A/B split yet
   showSurveyPopup?: boolean
@@ -69,7 +69,7 @@ export const initialState: UserState = {
   pairs: {},
   timestamp: currentTimestamp(),
   URLWarningVisible: true,
-  hideUniswapWalletBanner: false,
+  hideBaseWalletBanner: false,
   showSurveyPopup: undefined,
 }
 
@@ -98,8 +98,8 @@ const userSlice = createSlice({
     updateHideClosedPositions(state, action) {
       state.userHideClosedPositions = action.payload.userHideClosedPositions
     },
-    updateHideUniswapWalletBanner(state, action) {
-      state.hideUniswapWalletBanner = action.payload.hideUniswapWalletBanner
+    updateHideBaseWalletBanner(state, action) {
+      state.hideBaseWalletBanner = action.payload.hideBaseWalletBanner
     },
     updateDisabledUniswapX(state, action) {
       state.disabledUniswapX = action.payload.disabledUniswapX
@@ -214,7 +214,7 @@ export const {
   updateUserDeadline,
   updateUserLocale,
   updateUserSlippageTolerance,
-  updateHideUniswapWalletBanner,
+  updateHideBaseWalletBanner,
   updateDisabledUniswapX,
 } = userSlice.actions
 export default userSlice.reducer
